requirements.txt
streamlit
pandas
matplotlib
seaborn
pdfplumber
pinecone-client
google-generativeai
sentence-transformers
